Fixed typos and connectoin
A few items were fixed.  I added model from mongoose and fixed the typo model to module in User.js.

I added await to const userDoc due to it being an async function.

From here I tested the website and was able to check on MongoDB to make sure the data transfered over.