Update index.js
In this update :

I have defined a POST endpoint for my login.

I have created User Authentication and  I am using jwtwebtoken to create a token for the user.

I have also added error handling and a JSON response to the client.

I have also added a secret for the security of the JWT authentication