update index.js
First off I added fs and multer to the project.

Then I created a route to handle POST requests at the endpoint /post.   The server should expect a single file attachment with the field name 'file' processed by 'uploadMiddleware.sing('file') middleware.

After I recieve the file, the server will rename it to include its original extension and respons with the new file extension in JSON format.  This allows the client to display feedback to the user or store relevant info about the file that was uploaded.